Moved commands into correct folders
