+++ src/main/java/frc/robot/commands/DeployHatchCommand.java
@@ -17,7 +17,7 @@ public class DeployHatchCommand extends Command
         this.mode   = mode;
         this.deploy = deploy;
     }
-  
+
     // Called just before this Command runs the first time
     protected void initialize()
     {
@@ -29,7 +29,7 @@ public class DeployHatchCommand extends Command
         // Only execute this if we're in the correct mode.
         if(mode != Robot.oi.getMode())
             return;
-        
+
         if(deploy)
         {
             Robot.Lift.HatchPusher(true);
